requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.0
decorator==4.4.2
etelemetry==0.2.2
filelock==3.0.12
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
isodate==0.6.0
Keras-Preprocessing==1.1.2
-e git+https://github.com/ocr-senior-project/ocr.git@446c707590702c4b78f4336f7d89abd120d58fa5#egg=local_ocr_v2
lxml==4.5.2
Markdown==3.2.2
networkx==2.5
nibabel==3.1.1
nipype==1.5.1
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
prov==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pydotplus==2.0.2
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.1
pyxnat==1.3
rdflib==5.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
simplejson==3.17.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
traits==6.1.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
